extract binary helper functions out of NodeExecutionFunctions
