+++ packages/core/src/node-execute-functions.ts
@@ -16,26 +16,20 @@ import { ClientOAuth2 } from '@n8n/client-oauth2';
 import { Container } from '@n8n/di';
 import type { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 import axios from 'axios';
-import chardet from 'chardet';
 import crypto, { createHmac } from 'crypto';
-import FileType from 'file-type';
 import FormData from 'form-data';
 import { createReadStream } from 'fs';
 import { access as fsAccess, writeFile as fsWriteFile } from 'fs/promises';
 import { IncomingMessage } from 'http';
 import { Agent, type AgentOptions } from 'https';
-import iconv from 'iconv-lite';
 import get from 'lodash/get';
 import isEmpty from 'lodash/isEmpty';
 import merge from 'lodash/merge';
 import pick from 'lodash/pick';
-import { extension, lookup } from 'mime-types';
 import type {
-	BinaryHelperFunctions,
 	FileSystemHelperFunctions,
 	IAdditionalCredentialOptions,
 	IAllExecuteFunctions,
-	IBinaryData,
 	ICredentialDataDecryptedObject,
 	IDataObject,
 	IExecuteData,
@@ -50,13 +44,11 @@ import type {
 	IPollFunctions,
 	IRequestOptions,
 	IRunExecutionData,
-	ITaskDataConnections,
 	ITriggerFunctions,
 	IWebhookDescription,
 	IWorkflowDataProxyAdditionalKeys,
 	IWorkflowExecuteAdditionalData,
 	NodeExecutionWithMetadata,
-	NodeHelperFunctions,
 	NodeParameterValueType,
 	PaginationOptions,
 	RequestHelperFunctions,
@@ -80,7 +72,6 @@ import {
 	NodeOperationError,
 	NodeSslError,
 	deepCopy,
-	fileTypeFromMimeType,
 	isObjectEmpty,
 	ExecutionBaseError,
 	jsonParse,
@@ -96,9 +87,6 @@ import url, { URL, URLSearchParams } from 'url';
 
 import { Logger } from '@/logging/logger';
 
-import { BinaryDataService } from './binary-data/binary-data.service';
-import type { BinaryData } from './binary-data/types';
-import { binaryToBuffer } from './binary-data/utils';
 import {
 	BINARY_DATA_STORAGE_PATH,
 	BLOCK_FILE_ACCESS_TO_N8N_FILES,
@@ -111,6 +99,7 @@ import {
 import { DataDeduplicationService } from './data-deduplication-service';
 // eslint-disable-next-line import/no-cycle
 import {
+	binaryToString,
 	parseIncomingMessage,
 	parseRequestObject,
 	PollContext,
@@ -243,15 +232,6 @@ function digestAuthAxiosConfig(
 		}
 	}
 	return axiosConfig;
-}
-
-export async function binaryToString(body: Buffer | Readable, encoding?: string) {
-	if (!encoding && body instanceof IncomingMessage) {
-		parseIncomingMessage(body);
-		encoding = body.encoding;
-	}
-	const buffer = await binaryToBuffer(body);
-	return iconv.decode(buffer, encoding ?? 'utf-8');
 }
 
 export async function invokeAxios(
@@ -507,245 +487,6 @@ export async function httpRequest(
 	}
 
 	return result.data;
-}
-
-export function getBinaryPath(binaryDataId: string): string {
-	return Container.get(BinaryDataService).getPath(binaryDataId);
-}
-
-/**
- * Returns binary file metadata
- */
-export async function getBinaryMetadata(binaryDataId: string): Promise<BinaryData.Metadata> {
-	return await Container.get(BinaryDataService).getMetadata(binaryDataId);
-}
-
-/**
- * Returns binary file stream for piping
- */
-export async function getBinaryStream(binaryDataId: string, chunkSize?: number): Promise<Readable> {
-	return await Container.get(BinaryDataService).getAsStream(binaryDataId, chunkSize);
-}
-
-export function assertBinaryData(
-	inputData: ITaskDataConnections,
-	node: INode,
-	itemIndex: number,
-	propertyName: string,
-	inputIndex: number,
-): IBinaryData {
-	const binaryKeyData = inputData.main[inputIndex]![itemIndex].binary;
-	if (binaryKeyData === undefined) {
-		throw new NodeOperationError(
-			node,
-			`This operation expects the node's input data to contain a binary file '${propertyName}', but none was found [item ${itemIndex}]`,
-			{
-				itemIndex,
-				description: 'Make sure that the previous node outputs a binary file',
-			},
-		);
-	}
-
-	const binaryPropertyData = binaryKeyData[propertyName];
-	if (binaryPropertyData === undefined) {
-		throw new NodeOperationError(
-			node,
-			`The item has no binary field '${propertyName}' [item ${itemIndex}]`,
-			{
-				itemIndex,
-				description:
-					'Check that the parameter where you specified the input binary field name is correct, and that it matches a field in the binary input',
-			},
-		);
-	}
-
-	return binaryPropertyData;
-}
-
-/**
- * Returns binary data buffer for given item index and property name.
- */
-export async function getBinaryDataBuffer(
-	inputData: ITaskDataConnections,
-	itemIndex: number,
-	propertyName: string,
-	inputIndex: number,
-): Promise<Buffer> {
-	const binaryData = inputData.main[inputIndex]![itemIndex].binary![propertyName];
-	return await Container.get(BinaryDataService).getAsBuffer(binaryData);
-}
-
-export function detectBinaryEncoding(buffer: Buffer): string {
-	return chardet.detect(buffer) as string;
-}
-
-/**
- * Store an incoming IBinaryData & related buffer using the configured binary data manager.
- *
- * @export
- * @param {IBinaryData} binaryData
- * @param {Buffer | Readable} bufferOrStream
- * @returns {Promise<IBinaryData>}
- */
-export async function setBinaryDataBuffer(
-	binaryData: IBinaryData,
-	bufferOrStream: Buffer | Readable,
-	workflowId: string,
-	executionId: string,
-): Promise<IBinaryData> {
-	return await Container.get(BinaryDataService).store(
-		workflowId,
-		executionId,
-		bufferOrStream,
-		binaryData,
-	);
-}
-
-export async function copyBinaryFile(
-	workflowId: string,
-	executionId: string,
-	filePath: string,
-	fileName: string,
-	mimeType?: string,
-): Promise<IBinaryData> {
-	let fileExtension: string | undefined;
-	if (!mimeType) {
-		// If no mime type is given figure it out
-
-		if (filePath) {
-			// Use file path to guess mime type
-			const mimeTypeLookup = lookup(filePath);
-			if (mimeTypeLookup) {
-				mimeType = mimeTypeLookup;
-			}
-		}
-
-		if (!mimeType) {
-			// read the first bytes of the file to guess mime type
-			const fileTypeData = await FileType.fromFile(filePath);
-			if (fileTypeData) {
-				mimeType = fileTypeData.mime;
-				fileExtension = fileTypeData.ext;
-			}
-		}
-	}
-
-	if (!fileExtension && mimeType) {
-		fileExtension = extension(mimeType) || undefined;
-	}
-
-	if (!mimeType) {
-		// Fall back to text
-		mimeType = 'text/plain';
-	}
-
-	const returnData: IBinaryData = {
-		mimeType,
-		fileType: fileTypeFromMimeType(mimeType),
-		fileExtension,
-		data: '',
-	};
-
-	if (fileName) {
-		returnData.fileName = fileName;
-	} else if (filePath) {
-		returnData.fileName = path.parse(filePath).base;
-	}
-
-	return await Container.get(BinaryDataService).copyBinaryFile(
-		workflowId,
-		executionId,
-		returnData,
-		filePath,
-	);
-}
-
-/**
- * Takes a buffer and converts it into the format n8n uses. It encodes the binary data as
- * base64 and adds metadata.
- */
-// eslint-disable-next-line complexity
-export async function prepareBinaryData(
-	binaryData: Buffer | Readable,
-	executionId: string,
-	workflowId: string,
-	filePath?: string,
-	mimeType?: string,
-): Promise<IBinaryData> {
-	let fileExtension: string | undefined;
-	if (binaryData instanceof IncomingMessage) {
-		if (!filePath) {
-			try {
-				const { responseUrl } = binaryData;
-				filePath =
-					binaryData.contentDisposition?.filename ??
-					((responseUrl && new URL(responseUrl).pathname) ?? binaryData.req?.path)?.slice(1);
-			} catch {}
-		}
-		if (!mimeType) {
-			mimeType = binaryData.contentType;
-		}
-	}
-
-	if (!mimeType) {
-		// If no mime type is given figure it out
-
-		if (filePath) {
-			// Use file path to guess mime type
-			const mimeTypeLookup = lookup(filePath);
-			if (mimeTypeLookup) {
-				mimeType = mimeTypeLookup;
-			}
-		}
-
-		if (!mimeType) {
-			if (Buffer.isBuffer(binaryData)) {
-				// Use buffer to guess mime type
-				const fileTypeData = await FileType.fromBuffer(binaryData);
-				if (fileTypeData) {
-					mimeType = fileTypeData.mime;
-					fileExtension = fileTypeData.ext;
-				}
-			} else if (binaryData instanceof IncomingMessage) {
-				mimeType = binaryData.headers['content-type'];
-			} else {
-				// TODO: detect filetype from other kind of streams
-			}
-		}
-	}
-
-	if (!fileExtension && mimeType) {
-		fileExtension = extension(mimeType) || undefined;
-	}
-
-	if (!mimeType) {
-		// Fall back to text
-		mimeType = 'text/plain';
-	}
-
-	const returnData: IBinaryData = {
-		mimeType,
-		fileType: fileTypeFromMimeType(mimeType),
-		fileExtension,
-		data: '',
-	};
-
-	if (filePath) {
-		const filePathParts = path.parse(filePath);
-
-		if (filePathParts.dir !== '') {
-			returnData.directory = filePathParts.dir;
-		}
-		returnData.fileName = filePathParts.base;
-
-		// Remove the dot
-		const fileExtension = filePathParts.ext.slice(1);
-		if (fileExtension) {
-			returnData.fileExtension = fileExtension;
-		}
-	}
-
-	return await setBinaryDataBuffer(returnData, binaryData, workflowId, executionId);
 }
 
 export async function checkProcessedAndRecord(
@@ -1964,32 +1705,6 @@ export const getFileSystemHelperFunctions = (node: INode): FileSystemHelperFunct
 			});
 		}
 		return await fsWriteFile(filePath, content, { encoding: 'binary', flag });
-	},
-});
-
-export const getNodeHelperFunctions = (
-	{ executionId }: IWorkflowExecuteAdditionalData,
-	workflowId: string,
-): NodeHelperFunctions => ({
-	copyBinaryFile: async (filePath, fileName, mimeType) =>
-		await copyBinaryFile(workflowId, executionId!, filePath, fileName, mimeType),
-});
-
-export const getBinaryHelperFunctions = (
-	{ executionId }: IWorkflowExecuteAdditionalData,
-	workflowId: string,
-): BinaryHelperFunctions => ({
-	getBinaryPath,
-	getBinaryStream,
-	getBinaryMetadata,
-	binaryToBuffer,
-	binaryToString,
-	prepareBinaryData: async (binaryData, filePath, mimeType) =>
-		await prepareBinaryData(binaryData, executionId!, workflowId, filePath, mimeType),
-	setBinaryDataBuffer: async (data, binaryData) =>
-		await setBinaryDataBuffer(data, binaryData, workflowId, executionId!),
-	copyBinaryFile: async () => {
-		throw new ApplicationError('`copyBinaryFile` has been removed. Please upgrade this node.');
 	},
 });
